feat: add more translation
