+++ locales/ru/map.json
@@ -35,6 +35,7 @@
     "entryImage": "Изображение (необязательно)",
     "imageUrlExplanation": "рекомендуется: альбомный формат, мин. 400x300 пикселей, макс. 300 КБ",
     "imageUrl": "URL изображения",
+    "tutorialModalTitle": "Как получить URL-адрес изображения",
     "InfoHowToCopyURL": "Вы можете легко получить URL-адрес любого изображения в сети, просто щелкнув по нему, чтобы открыть контекстное меню и выбрать «Копировать URL картинки». Вставьте этот URL (который будет заканчиваться на .jpg, .png или подобное) в эту строку! ",
     "imageLink": "Ссылка (при нажатии на изображение, необязательно)",
     "cancel": "отменить",
